+++ crates/captval/src/secret.rs
@@ -30,8 +30,9 @@ impl Secret {
     }
 }
 
+#[cfg(feature = "hcaptcha")]
 #[cfg(test)]
-mod tests {
+mod hcaptcha_tests {
     use super::Secret;
     use crate::Code;
     use crate::Error;
@@ -45,19 +46,6 @@ mod tests {
     #[test]
     fn empty_string_is_rejected() {
         let secret = "".to_string();
-        assert_err!(Secret::parse(secret));
-    }
-    #[cfg(feature = "ext")]
-    #[test]
-    fn secret_longer_than_secret_len_is_rejected() {
-        let secret = "1234567890123456789012345678901234567890123".to_string();
-        assert_err!(Secret::parse(secret));
-    }
-
-    #[cfg(feature = "ext")]
-    #[test]
-    fn secret_shorter_than_secret_len_is_rejected() {
-        let secret = "12345678901234567890123456789012345678901".to_string();
         assert_err!(Secret::parse(secret));
     }
 
@@ -79,6 +67,45 @@ mod tests {
     #[test]
     fn test_v2_secret_key_is_valid() {
         let secret = "ES_a5e0b5406e2b4c939ce48946389463894638473b1c".to_string();
+        assert_ok!(Secret::parse(secret));
+    }
+}
+
+#[cfg(feature = "recaptcha")]
+#[cfg(test)]
+mod recaptcha_tests {
+    use crate::{Code, Error, Secret};
+    use claims::{assert_err, assert_ok};
+
+    #[test]
+    fn whitespace_only_secrets_are_rejected() {
+        let secret = " ".to_string();
+        assert_err!(Secret::parse(secret));
+    }
+
+    #[test]
+    fn empty_string_is_rejected() {
+        let secret = "".to_string();
+        assert_err!(Secret::parse(secret));
+    }
+
+    #[test]
+    fn error_set_contains_missing_secret_error() {
+        let secret = "".to_string();
+        if let Err(Error::Codes(hs)) = Secret::parse(secret) {
+            assert!(hs.contains(&Code::MissingSecret));
+        }
+    }
+
+    #[test]
+    fn test_valid_secret_key() {
+        let secret = "6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI".to_string();
+        assert_ok!(Secret::parse(secret));
+    }
+
+    #[test]
+    fn test_another_valid_secret_key() {
+        let secret = "6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe".to_string();
         assert_ok!(Secret::parse(secret));
     }
 }