+++ crates/captval/tests/ui/recaptcha_pass/test_complete_success_response.rs
@@ -1,7 +1,8 @@
 mod helper;
 
+use captval::Captval;
 use chrono::{TimeDelta, Utc};
-use captval::{Code, Captval};
+use claims::assert_ok;
 use serde_json::json;
 use wiremock::matchers::{body_string, method, path};
 use wiremock::{Mock, MockServer, ResponseTemplate};
@@ -22,7 +23,7 @@ async fn main() {
     let token = helper::random_string(100);
     let remoteip = mockd::internet::ipv4_address();
     let sitekey = mockd::unique::uuid_v4();
-    let secret = "ES_InvalidSecretString".to_string();
+    let secret = format!("0x{}", hex::encode(helper::random_string(20)));
 
     let expected_body = format!(
         "response={}&remoteip={}&sitekey={}&secret={}",
@@ -57,16 +58,8 @@ async fn main() {
     };
     let response = form.valid_response(&secret, Some(uri)).await;
 
-    claims::assert_err!(&response);
-
-    if let Err(codes) = response {
-        match codes {
-            captval::Error::Codes(hash_set) => {
-                assert_eq!(hash_set.len(), 2);
-                assert!(hash_set.contains(&Code::InvalidSecretExtNotHex));
-                assert!(hash_set.contains(&Code::InvalidSecretExtWrongLen));
-            }
-            _ => unreachable!(),
-        }
-    };
+    assert_ok!(&response);
+    let response = response.unwrap();
+    assert!(&response.success());
+    assert_eq!(&response.timestamp().unwrap(), &timestamp);
 }